refactor(core): extract bootstrap helpers to core/game_bootstrap.js and use them from game.js
